Add temporary fix for unresolved/unreduced type-families.
Some type families reduce to type integer-literals (such as in
the case of BitSize), but they are currently not dealt with.
Where a solution is either to remove them in the GHC->Core
conversion. Or to add type-function resolution to CLaSH.

For now we just add a dirty hack for the BitSize type-family,
and its Unsigned and Signed instances.