Retype free variables after ghc cast removal
When transforming GHC's Core to CLaSH's Core language, we drop all
casts, as they don't exist within CLaSH's core.

Casts often happen in two places, at the definition site of a global
binders, and at the use sites of global binders.

Given a situation where:
- 'f's body has type A, but is casted to B
- 'f's use sites have the references to f casted from B to A
removing the casts poses a problem, because the use sites of
'f' still think it has type B. Hence we have to update the
types at all use sites of 'f' back to 'A', and all is right
with the world.