Compile 'Fixed.hs' with -O0 so that we don't _the_ strange Id
Compiling with -O (default), gives rise to an identifier names
'$fNumFixed', which is marked as AlwaysInline, but has no
Unfolding. CLaSH hence considers it a BlackBox and complains
it cannot find a corresponding template.

Compiling with -O0 makes _the_ strange Id disappear from the
interface file, and all is well.