Add exception to the inlining limiter for dictionaries that are known to be benign
inlineNonRep normally only allows inlining of top-level entities
100 times within a function body. This is to prevent the compilier
from going into an infinite inlining loop on recursive, and als not
representable values.

The problem is that this inlining limit is easily hit when creating
vectors with large amount of integer values. As these are all of
the form 'fromInteger $dNum lit'. Meaning that on the 100th inlining
of the Num dictionary, the compiler complains and exits.

By adding exemptions for dictionaries that are benign, such as
the Num dictionaries of Unsigned and Signed, we can properly deal
with large vectors of numbers. Without having to worry that we'll
spin the compiler in an infinite loop.