Don't strip args in flattenNode when they are mentioned
stripArgs erroneously removed args/lambda's when the argument
wasn't equal to one of the lambda's. As a result, it was
introducing free variables.  It did this under the
assumption that normalised Core is in ANF.

However, because of higher-order primitives, fuction-typed
arguments might not be in ANF. Meaning that a partially
applied function might refer to one of the lamba's.

The fix is: instead of checking that the argument
was not one of the lambda's, to check if the free variables
of the argument is not one of the lambda's.