Add ~COMPNAME labels to memory elements
For example, when a blockRam blackbox is instantiated it gets the
label: blockRam_~COMPNAME_~SYM[0]
When synthesis tools report longest combinational paths, they
use these labels to indicate where a contributing signal originates.
By adding COMPNAME to the label, it will be easier to determine
in which function the blockRam was used, making it easier to
determine which part of our CLaSH design is contributing to the
longest path.