Treat bottoming NoUnfolding as 'undefined'
The CLaSH/GHC frontend inspects .hi interface files to look for
the CoreExpr definitions of compiled/installed functions. When
it doesn't find one, or finds one that is specifically marked as
NOINLINE, it expects a BlackBox to fill in the "hole" during
code-generation.

This is the reason why the 'clash-prelude' is compiled with the
'-fexpose-all-unfoldings', so that all CoreExprs are put in the
.hi interface files for the CLaSH compiler to translate. However,
expressions that are guaranteed to be bottom values, are still
not given an Unfolding in the interface files. They are
explicitly given a 'NoUnfolding'.

Now, when CLaSH encounters a bottoming value with a 'NoUnfolding',
it simply replaces it with 'undefined'. Since we only deal with
the pure (exception-free) subset of Haskell, all bottoms are
equivalent, so it is just fine to use 'undefined'.

Ultimately, this stops CLaSH complaining/erroring out on expressions
which we don't care about anyway.