Re-enable strictness analysis
This used to be disabled because it enabled GHC to create a
worker/wrapper which exposed the Signal constructor, which in
turn broke compilation.

We now prevent this by turning off strictness analysis in the
"CLaSH.Signal.Internal" module, so that functions applying
the Signal constructor do not get a strictness/demand/CPR
annotion. This in turn prevents GHC from creating a worker/wrapper
when these functions are used.