Lift join-points instead of inlining them
Join points are created in complex-ish case-statements by GHC.
They can lead to exponentional blow-up in compilation time.

This in exhibited in issue #70, and this patch gets us part
of the way of fixing it. Compilation time has now reach an
acceptible lvl (e.g. 2-3 secs) instead of needing to kill the
machine because it runs out of memory.

As a consequence, CLaSH generates many new HDL files with weird
names that never occured in the user program. A follow up patch
will have to fix that.