Record provenance in lifted binders
For now, we just prefix the name of the function from which
a lifted binder originated, to the name of the new global
binder. At least this reflects the provenance of certain
"weird" components which sometimes appear because of the
way GHC compiles pattern matching functions to case-statements;
such as what happens in issue #70.

Fixes #70