Always check if we hit the specialisation limit.
Before, we only checked if we hit a specialisation limit if we
were specialising on non-representable terms, but left
specialisations on types and (representable) constants unchecked.

It's possible to create pathological cases, which are
non-sensible from a circuit perspective, which would throw the
compiler in an endless specialisation loop.

So now, for ALL specialisations, we check if we are hitting the
specialisation limit.