Pull reset low afer 1 simulation step in verilog
Before, we would pull the reset low when simulation started.
In some instances, where the reset value of a register was not
a "true" constant, but a constant coming from outside of the
Verilog module, the register would get all 'z' values.

As to why no all constants get propagated: there are reasons...