"Fix" 'Eq' instance of 'Vec'
Equality (==) was treated as a primitive by the CLaSH compiler.
While its Haskell definition was just fine, the CLaSH compiler
translated it directly to the HDL's equality operator. While
this was just fine for most element types of a Vector, this
translation was not meaning-preserving for element-types where
the element's equality in Haskell was different from the HDL's
equality.

Hence, we stop treating Vector equality as a primitive.