DEC: extract disjoint expresions from alternatives before the subject
In a rush to fix issue #88, i first checked for disjoin
expressions in the subject of a case-expression, before checking
the alternatives of a case-expression. Any expression seen in
the subject, would considered not to be disjoint from the
expression in the alternatives.

However, given a case-expression of the form:

> case (f ..) of A -> (f ..); B -> (f ..); C -> (f ..)

I would not be able to lift out the 'f's in the alternatives
because they are not disjoint from the 'f' in the subject.

So, by checking the alternatives first, I can lift out all the
'f's in the alternatives, but leave the 'f' in the subject
(because it is not disjoint from the 'f's in the alternatives).
Hence DEC now has more oppertunities to make the circuit smaller
while still being correctly implemented.