Stop reversing alts list in netlist generation
We used to do this because GHC puts the default alternative in
the first position. Simply reversing the list of alternatives
is would make VHDL/Verilog be the same as the original Haskell,
under the assumption that all patterns are unique. This however
isn't true for literal patterns, where the scrutinee might not
have enough bits to represent all te alternatives. So what
we do now is just put the Default pattern at the end.

Fixes #91