DEC: root of the case-tree must contain at least 2 alternatives
Before, a case-tree was disjoint when it has more that 2 branches,
or, if it has a single branch which is itself disjoint. We do
this so we can look through pattern matches on product types to
find disjoin applications. However, we must _not_ do this for the
root of the case-tree.

Fixes #92