etaTL: move lambda's through let-bindings
After moving etaTL after application propagation (in order to preserve
more Haskell names), we sometimes ended up with a lamba in the body
of a top-level let-binding, where the argument binder of the
enclosing lambda was not propagated because we moved eta-expansion
after application propagation.

What we do now is ensure that all (created) lamba's in the body
of a let-binder are move upward.

Fixes #132