DEC: don't throw away transformed arguments.
When an argument of a function that isn't disjoined is being
substituted (because it is disjoint), CLaSH's DEC transformation
used the old argument instead of the new argument.

Consequently, the DEC transformation didn't make any changes but
still marked the transformation as a success. This threw CLaSH
into an endless loop sometimes.

Fixes #140