Bind/Inline join-points that are simple void wrappers
Normally, we do not inline join points because it doubles
compilation time. However, when these join points are of the form:

> \(w :: Void) -> f a b c

i.e. a wrapper around a function 'f' where 'w' is not used we do
inline them. We do this because after tuning of Worker/Wrapper
transformations in GHC, our CSE pass missed a few oppertunities
which in some case resulted in circuits that were twice as big.

Inline join-points that are void wrappers does not increase
compilation time (in any significant way)