"Fix" the Eq instance of Vec
GHC was still generating Core that was pattern-matching on `Cons`,
which CLaSH cannot deal with. So we now use `length` to check the
length of the vector we are comparing for equality.

Sadly, this adds a `KnownNat` constraint, I hope this doesn't
break code in the wild. Otherwise I'll have to up the version
number of the clash-prelude