DEC: store substitution environment
Applied functions are only lifted when they are considered
disjoint. Function are disjoint when they only occur once
(up til that point) in a case-branch.

DEC creates let-bindings for these lifted functions. However,
because DEC lifts multiple functions at once, not only must
substitions happen in the original case-expression, substitutions
must also happen in the lifted let-bindings.

However, the substitution mechanism uses the same mechanism
as the disjoint-check, i.e. substitution only takes place when
a disjoint function has not been seen up til that point.

Previously, lifted out function were substituted on with an
empty "list-of-already-seen-functions", which does not match
the "list-of-already-seen-functions" when they were marked
for disjointess. Consequently, bogus substitutions were occuring,
as the traversal mechanism was initialized with an empty list.

As a consequence of this bug, functions which were originally
not considered disjoint in one of the other lifted functions,
were now being bogusly substituted by the results of other
lifted functions.

To fix this, lifted function are annotated with the
"list-of-already-seen-functions" at the point they were lifted
from. This means that the substitution mechanism now works
as it is supposed to.