Simplify KnownNat constraints in for Numeric type operations
Currently, this slows down the performance of the operations
because wrap-around points are computed inside of the operation,
and are not specialised due to the NOINLINE pragmas.

In a planned patch, these NOINLINE pragma's will be removed when
the clash-prelude is compiled for fast compilation, and the
performace lost by the current patch will be recovered.