Remove partially implemented instances on Signal
And deprecate all f1 functions that can easily be written as:

`fmap f` or `liftA2 f`

Completely implemented instances of classes which contain
operator methods, such as the `Num` and `Fractional` class are
kept... for now.
We do this because lifting operators in an infix position is
difficult.

Perhaps, for consistency, these instances will removed at a later
date.