Remove source strictness annotations on data types
Strictness annotations introduce casts between Integer and
CLaSH' numeric primitives (e.g. Unsigned, Index, etc.), and
CLaSH will error when it sees such casts. The reason being
that Integer is a completely unconstrained type and is
currently (erroneously) to a 64-bit signed number in the
generate HDL. This might however mean that we lose bits if
the numeric type had more bits than 64.