Support "primitive directory" annotations
This means that primitive definitions can be bundled with libraries
and CLaSH can find them.

i.e. give a foobar.cabal file with:

> data-files: primitives/vhdl/Foo.json
>             primitives/verilog/Foo.json
>             primitives/systemverilog/Foo.json

and a Foo.hs file:

> module Foo where
>
> import qualified Paths_foobar
> import           System.FilePath
> import           System.IO.Unsafe
>
> {-# ANN module (Primitive VHDL (unsafePerformIO Paths.getDataDir </> "primitives </> "vhdl")) #-}
> {-# ANN module (Primitive Verilog (unsafePerformIO Paths.getDataDir </> "primitives </> "verilog")) #-}
> {-# ANN module (Primitive SystemVerilog (unsafePerformIO Paths.getDataDir </> "primitives </> "systemverilog")) #-}

the CLaSH compiler can figure out that the BlackBox definitions
corresponding to the primitives defined in Foo.hs can be found
in the `datadir` directories of the installed `foobar` package.