Add clock ports of cached results
Clash caches `Term -> Component` translation; also, it uses a
writer monad to keep track of the implicitly used clocks.

However, when a cached result was used, the monadic action that
created that result was, of course, not run. This meant that
the clocks for that sub-component werent added to the current
writer monad session.