Verilog: truncate Integer when converting to sized
Integer often winds up being represented by 64 bits.  When using
fromInteger# to assign to, say, a 'BitVector 8', this is fine in
Verilog, because the top 64-8 bits get discarded.

But because the template is an expression template, it can also get
applied inside of a bitstring concatentation like

    {x, y, $unsigned(an_integer), z}

Here the difference between an 8-bit and a 64-bit value changes the
meaning of the expression.

This commit explicitly slices Integers when moving into four common
sized integer-like types.

See #219 (this may not be a complete fix for that bug) and #220 (filed
as a result of this work)