Verilog: don't generate intermediate 'reg' signals
Instead of assigning the result of a declaration needing a
'reg' signal to an intermediate signal, we assign to the result
directly. This is made possible by declaring the result as a
'reg' (we used to default everything to 'wire'), depending
on whether the declaration is a conditional assignment, or
the used BlackBox says that its result must be a 'reg'.