Increase reset pulse duration generated in testbenches by 100x
In the old clash the reset pulse was 2ns long and clock started after 3ns.
Somewhere this got changed to 20ps and 30ps.
But such extremely sort reset pulse cause problems with the simulation models of the Xilinx DDR ip cores.

So change the reset pulse duration back to their old values.

Signed-off-by: Christiaan Baaij <christiaan.baaij@gmail.com>