Generate hierarchical component names
Component names are now the fully qualified names of functions,
i.e. they include the (hierarchical) module name; however,
instead of periods `.`, we use underscores `_` to seperate
the parts. We need to do this because a period is not a valid
character of a basic identifier; for both VHDL and Verilog.