VHDL: delay result of memory elements by a small amount
This is to prevent issues with delta-cycles, making it seem
as if we are working with multiple clocks. The problem arises
when clocks are put into tuples, and then extracted, adding
1 delta cycle delay. Consequently, it will seem as if one
register connected directly to a clock, and one register
connected to a packed-into-tuple, and then unpacked-from-tuple,
clock is connected to a completely different clock. Thus
the 1-delta-cycle delayed register is seeing values from the
new state, in simulation, while the synthesis behaviour will be
it seeing things from the old state.

By delaying the output of state elements (registers, blockRAM,
etc.) by 1ps, we do not get this annoying delta-cycle behaviour.