VHDL: add Product type to cache if its name is queried
Sometimes we only construct a product, and dont use it in
any other way (i.e. we convert it to an SLV right away).
Then only the constructor is called, and hence only
`tyName` is queried. In this case we need to add it to
the cache so that proper type declarations and conversion
functions are made.