Evaluator: create new literals with their proper types
When creating new literals (as result of compile-time evaluation),
we used to recreate their the type for each new literal created.

But because types are internally unique this results in errors like:
    CLaSH.Rewrite.Util(102): Error when applying rewrite evalConst to:
    [...]
    Changes type from:
    CLaSH.Sized.Internal.Signed.Signed 8
    to:
    CLaSH.Sized.Internal.Signed.Signed 8

Now we given them their proper type by looking it up in the environment
where the original value was.

Signed-off-by: Christiaan Baaij <christiaan.baaij@gmail.com>