Projections bind to a variable when a scrutinee of a case
In VHDL, subjects/scrutinees of a conditional assignment must
have a statically known range. However, there are cases where
we project e.g. an `unsigned` out of a Sum-of-Products type
and we convert the SLV using `unsigned(..)`. At that time
the projection no longer has a "statically known range" because
`unsigned` is unconstrained.

To prevent this from occuring, we assign projections to a
signal/wire when they are a scrutinee/subject of a
case-expression.