Stop binding primitives to signals when arguments of a SoP DataCon
Sum-of-Products (SoP) are encoded as SLV/BitVectors, and so all
the arguments must be converted to SLVs. In previous iterations
of the Clash compiler, the toSLV conversion process was very
limited, meaning basically that we needed all arguments of an
SoP DataCon to be variables.

Clash now has full toSLV support meaning we shouldn't bind
arguments to wires/signals when we don't have to.

This further reduces the number of extra internal signal/wires
that Clash would've otherwise generated.