Inline function generated as part of the normalisation process
During normalistation, let-bindings with a function type are
sometimes lifted to list of top-level bindings. We do this
for termination aspects, performance aspects, CSE and DEC
opportunities, etc.

However, to ensure that the generated HDL has a hierarchy
similar to the original Haskell source, we inline these
generated functions again at the end of the normalisation
pipeline.