Always inline single-binder normalised functions.
Before we only inlined them when the let-bound expression used
the arguments in the same order as the encompassing lambda.
However, to the intention is to just have the "major"
hierarchical structure present in the generated VHDL. So all
small wrappers should be inlined as part of the cleanup process.