Evaluator: support rounding of constants with certain types
(To use this run clash with the option: -clash-float-support)

This supports round,ceiling,floor,truncate on constants of type Float or Double to
 * Signed n
 * Unsigned n
 * Word

ceiling,floor,truncate also work to Int and Integer.
But not round to Int or Integer doesn't work (yet).

This is because GHC internally uses the FFI to do this rounding in C.