Change type of TopEntity.t_outputs (#121)
Haskell functions can only ever return a single value, meaning that t_outputs
taking a list of PortNames is wrong: any name after the first one is simply
thrown away and forgotten. This is bug prone and easy to misunderstand, for
example when you return multiple values with a tuple.

Instead, change the type to PortName, making it clear (at the type level) that
naming an output requires either a single name, or a PortField with a list of
names (for example, to name components of an output tuple).

This changes the type of t_outputs *and* its name to match this fact. Output
names are still auto-generated by default.

Signed-off-by: Austin Seipp <aseipp@pobox.com>