'inlineClosed' is now called 'inlineWorkFree'
A closed expression is an expression with no free variables.
The intention of the 'inlineClosed' expression was to
inline the body of a function which was applied to arguments
that were closed, knowing that it would reduce to a
constant. It did this in a roundabout way, by checking
whether an argument was a constant (literal, data
constructor, or primitive), which was not very satisfying.

However, being "closed" isn't enough to be certain that
an expression can reduce to a constant. It should also
not be recursively defined. Feedback loops are expressed
by recursive let-bindings, where the binding will have a
Signal type. So we also checked whether the expression
did not have a Signal type.

So what we actually wanted is to inline a function body
when we knew that the entire expression was going to be
work-free, i.e. not contribute to the size of the circuit.
Hence the name-changed to 'inlineWorkFree'.

Also, instead of checking whether the arguments of a
function are "constants", we now actually check whether
they have no free variables and are not of a Signal type.

In addition to that, inlineWorkFree now also inlines
functions that are marked NOINLINE because  inlining a
work-free expression will not affect the performance of
the circuit.