Change names so it's clearer that things are clock generators
It is especially confusing that in old-style clash, `systemClock`
was used to connect a circuit to the system clock, whilst now it
is a clock generator for the sytem domain. To make matters worse
the clock generator synthsizes correctly, but for HDL synthesis
results in a clock line that is connected to ground.