Evaluator: Fix incorrect heap short-cutting
Previously we had an optimization to avoid pointer-chasing due to trivial
bindings in the heap. Namely, when we were asked to insert a binding with
rhs (Var _ v2) into the heap, we would instead return the heap
unchanged and instead just return the binder v2. However, what if v2 isn't
present in the heap (TODO: when could this happen?)

Well, then things blew up. We now instead perform a slightly less aggressive
optimization: when we insert a binding with RHS (Var _ v1) into heap [v1 = v2],
we return the heap [new_bndr = Var _ v2, v1 = v2].