Add primitive warnings
This commits adds the possibility to specify warnings in the clash
primitive blackbox declarations. These will be shown to the user when
generating a NetList using such primitives.

A clash flag `-fclash-no-prim-warn` will remove the warning (which are
enabled by default). Maybe these warnings shouldn't be shown by default?

The main use-case for these warnings is to warn the user that the
primitive they are directly or indirectly using is not synthesizable.
It is obviously fine to use such primitives in a `testbench` (and maybe the
warnings should actually be disabled in testbenches).

The only blackboxes that currently have the warnings enabled are:

- `clockGen`
- `tbClockGen`

I will add warnings to more primitives as I think about them. I welcome any
suggestions to improve this PR.

My own suggestions for improvements (I will make further commits to
address them if they are of interest to the Clash contributors):

- Print warning to `stderr` instead of `stdout`
- Coloured warnings (orange?)
- Disable warnings for the `testbench` netlist generation
- Remove duplicate warnings (for the same primitive)