pack/unpack should just be wires
When using a custom representation the HDL type is already a bitvector,
so pack and unpack don't have to do anything.

We accomplish this by applying the (generated) pack/unpack function with
   dontApplyInHDL :: (a -> b) -> a -> b
This function applies it first argument to it's second argument.
But only in Haskell, in HDL it just returns it's second argument unmodified.