Faster capture-free substitution
This patch removes the use of the `unbound-generics` for names
and bound-variables, as profiling indicated that at least half
of the time of the normalisation pass was spend inside functions
concerning themselves with keep invariants alive for the
substitution functions of `unbound-generics`.

In stead we implement our own substitution, alpha-equivalence,
and free variable calculation; and use the capture-free
substitution method referred to as "The rapier" in section 3.2 of

      "Secrets of the GHC inliner"

Since we change the representation for names, this patch
touches a lot of modules, which cannot be split into multiple
commits.

As a result of this patch, we get a 10x to sometimes 50x speedup
on the normalisation pass; which is the bulk of the work done
by the Clash compiler.