Remove most `'length`/`'range` constructs in `Clash_Sized_Vector.json`
Referencing `'length` and `'range' might cause problems when dealing
with 'empty' vectors, such as a vector of units. Thus, this commit
removes references to it, except in cases where it is used on ~RESULT,
as we assume that a function result is not empty (as it should have been
removed by Clash). This commit only affects VHDL, as we do not use
these constructs for Verilog.