Make ANF shadowing-safe
Before, ANF _could_ pick new binders that would shadow existing
bound variables, and would thus end up with a let-expression with
two (or more) equally named binders.

We now have 'freshenTm' return an InScopeSet that additionally
contains all the bound variables of the expression it just made
unique. This gives ANF a starting InScopeSet that contains both
the free and bound variables of the expression. So whenever ANF
creates a new binder, it will not shadow an existing bound
variable (or in-scope free variable of course).

Additionally we fix a fault where for ANF for a case expression
with multiple alternatives could have picked conflicting new
names for the binders that would bind the alternative expressions.