removedArg @ Clock/Reset is a normal constant (#442)
Normally constants with a Clock or Reset type are lifted out so
we can do CSE, so that we only get 1 clock generator, and 1
reset generator. We do this by checking the type of the constant.

However, if the constant is actually a `removedArg`, introduced
to replace arguments of primitives that are unused, we should
leave them where they are.