Stop Evaluator introducing free variables
Compile-time evaluation of `splitAt` and `fold` could potentially
introduce free variables by using the parts of the type of the
Primitive to construct other types.

What it should've been doing, and is doing now, is applying the
type of the primitive to the types applied to the primitive, thus
beta-reducing the type-variables, get getting out types not
mentioning any free variables.