Add GADT support for Vec-like constructs
This commit adds support for Vec-like GADT constructs, as well as
simpler ones. To understand what changed, let's consider a small
example:

    data T a where
      SBool :: Bool -> T Bool
      SInt  :: Int  -> T Int

I.e., a sum-of-products with two constructors: `SBool` and `SInt`. When
`T a` is constructed with the constructor `SBool`, we're sure that
`a ~ Bool`. Likewise, whenever `T a` is constructed with `SInt`, we know
that `a ~ Int`. The inverse is true too: whenever we match on each of
the constructors we gain type information. Consider:

    f :: T a -> a
    f t =
      case t of
        SBool b -> not b
        SInt i  -> i + 1

At first it might seem that GHC would reject this program: the two
alternatives yield wildly different types. But, as alluded to earlier,
the pattern match introduces some extra information:

    f :: forall a . T a -> a
    f t =
      case t of
        SBool {dt: a ~ Bool} b -> (not (b |>dt)) |>sym dt
        SInt  {dt: a ~ Int}  i -> ((i |>dt) + 1) |>sym dt

GHC inserts a proof (inside curly braces) that `a ~ Bool` when matching
on `SBool`, and a similar proof for `SInt`. On the right hand side of
each alternative it casts the polymorphic `a` to `Bool`/`Int` with
`b |>dt`/`i |>dt` respectively. Thanks to the given proof (which was
obtained during construction), this typechecks. Because the type of the
case has to be `a`, a symmetric cast is applied with `|>sym dt`.

So how do we use this information to compile seemingly polymorphic
functions? Well, consider what happens when `f` is specialized such that
it is monomorphic (like every sound Clash circuit). That is:

    f @Int

The function definition now looks like:

    f :: T Int -> Int
    f t =
      case t of
        SBool {dt: Int ~ Bool} b -> (not (b |>dt)) |>sym dt
        SInt  {dt: Int ~ Int}  i -> ((i |>dt) + 1) |>sym dt

Clearly, `Int ~ Bool` is absurd! We can therefore remove it (see
transformation `caseElemNonReachable`) leaving us with:

    f :: forall Int . T Int -> Int
    f t =
      case t of
        SInt  {dt: Int ~ Int}  i -> ((i |>dt) + 1) |>sym dt

..a case with a single alternative.

This is not enough to support `Vec`-like GADTs though. Why not? Well,
let's look at the definition of `Vec`:

    data Vec :: Nat -> Type -> Type where
      Nil  :: Vec 0 a
      (:>) :: a -> Vec n a -> Vec (n + 1) a

Similar to our previous function `f`, we can build a function pattern
matching on `Vec`:

    g :: Vec n Bool -> Bool
    g v =
      case v of
        Nil       -> False
        Cons x xs -> True

What would the annotated version look like? Well..

    g :: forall n . Vec n Bool -> Bool
    g v =
      case v of
        Nil  {dt: n ~ 0}                                     -> False
        Cons {n1:Nat} {dt: n~n1+1} (x :: a) (xs :: Vec n1 a) -> True

Alright, so let's do the same trick again: specialize `g`.

    g @0

By filling in the `n`s we get:

    g :: Vec 0 Bool -> Bool
    g v =
      case v of
        Nil  {dt: 0 ~ 0}                                     -> False
        Cons {n1:Nat} {dt: 0~n1+1} (x :: a) (xs :: Vec n1 a) -> True

So what gives? The alternative `Cons` is not _clearly_ absurd now. Of
course, intuition dictates that if `0 ~ n1+1` then `n1 ~ 0-1 ~ -1`. As
we're operating in the domain of naturals, this _is_ absurd! Our
intuition relied the fact that `-` is the inverse of `+`. This
property (+injectivity) is not being tracked by GHC/Haskell/Clash
though, so we can't really do anything in general. This commit adds
hardcoded logic for `+`, see `elemExistentials`.

Future work: make "polymorphic" GADTs synthesizable by treating them
(in as sense) as sum-of-product types.