Stop ANF from let-binding a projected pattern variable
This could induce an exponential blow-up in the number of
transformation. Also, then intended normal form is still
reached.

Aside from the change to ANF, we also needed a change to
inlineCleanup. We'll illustrate using an example as to
why:

GHC sometimes transforms:

```
f clk rst = let s = register clk rst (s + 1) in s
```

to:

```
f clk rst = register clk (f clk rst + 1)
```

This is problametic for Clash because we go from a recursive
let-binding, which can be translated to a feedback loop, to
a recursive function, which describes an infinitely large
circuit.

So what was happening before the change with ANF? Well, we
started out with:

```
f = \ds ->
 let clk = case ds of (sel0,_) -> sel0
     rst = case ds of (_,sel1) -> sel1
 in  register clk rst 0 (f (clk,rst) + 1)
```

which ANF then transformed to:

```
f = \ds ->
  let caseAlt0 = sel0
      sel0 = case ds of (sel2,_) -> sel2
      clk  = caseAlt0
      caseAlt1 = sel1
      sel1 = case ds of (_,sel3) -> sel3
      rst  = caseAlt1
      appArg0 = (clk,rst)
      appArg1 = f appArg0
      appArg2 = appArg1 + 1
  in  register clk rst 0 appArg2
```

then after `bindConstantVar` and `topLet` we get:

```
f = \ds ->
  let sel0 = case ds of (sel2,_) -> sel2
      sel1 = case ds of (_,sel3) -> sel3
      appArg0 = (sel0,sel1)
      appArg1 = f appArg0
      appArg2 = appArg1 + 1
      result  = register clk rst 0 appArg2
  in  result
```

now comes `inlineCleanup`, which among other things inlines
let-bound projections, in the let-binder isn't a user-defined
name:

```
f = \ds ->
  let appArg1 = f ( case ds of (sel2,_) -> sel2
                  , case ds of (_,sel3) -> sel3 )
      appArg2 = appArg1 + 1
      result  = register clk rst 0 appArg2
  in  result
```

and then `recToLetRec` changes it to:

```
f = \ds ->
  let appArg1 = result
      appArg2 = appArg1 + 1
      result  = register clk rst 0 AppArg2
  in  result
```

and we have our feedback-loop again. So what happened when we
changed the behaviour of ANF? Well, ANF then resulted in:

```
f = \ds ->
  let sel0 = case ds of (sel2,_) -> sel2
      clk  = case ds of (sel0,_) -> sel0
      sel1 = case ds of (_,sel3) -> sel3
      rst  = case ds of (_,sel1) -> sel1
      appArg0 = (clk,rst)
      appArg1 = f appArg0
      appArg2 = appArg1 + 1
  in  register clk rst 0 appArg2
```

then, `deadCodeElim`, `bindConstantVar`, and `topLet` later we
end up with:

```
f = \ds ->
  let clk = case ds of (sel0,_) -> sel0
      rst = case ds of (_,sel1) -> sel1
      appArg0 = (clk,rst)
      appArg1 = f appArg0
      appArg2 = appArg1 + 1
      result  = register clk rst 0 appArg2
  in  result
```

and now `inlineCleanup` cannot inline the `clk` and `rst`
binders anymore because they have "user-defined" names, so
we end up with:

```
f = \ds ->
  let clk = case ds of (sel0,_) -> sel0
      rst = case ds of (_,sel1) -> sel1
      appArg1 = f (clk,rst)
      appArg2 = appArg1 + 1
      result  = register clk 0 appArg2
  in  result
```

And the issue now is that `recToLetRec` isn't sufficiently
smart enough to see that `f` is applied to unaltered arguments.
i.e. it could see the isomorphism of:

ds ~ (case ds of (sel0,_) -> sel0,case ds of (_,sel1) -> sel1)

but not of:

ds ~ (clk,rst)

where `clk` and `rst` are same projections as above, but simply
let-bound.

Now, instead of making `recToLetRec` a "whole lot" smarter, we
simply allow `inlineCleanUp` to inline projections, even if the
let-binder has a "user-defined" identifier. Why? because the
identifiers `clk` and `rst` weren't actually written by the
user at all! They were the result of the desugaring of the
implicit parameters that we use to implicitly pass along hidden
clocks and resets.

So, there's room for improvement still:

1. Make `recToLetRec` smarter
2. Ensure that identifiers marked as "User" by GHC, but are
   clearly invalid Haskell identifier (e.g. $dIP123), are stripped
   from their "User" marking by Clash so that Clash can more
   freely inline them. Where the idea is that Clash tries to
   preserve as many user-defined identifiers in the generated HDL
   as possible.