Add template function for `GHC.Word.WX#` data constructors
Verilog does not support truncating on arbitrary expressions. Instead,
users need to use assign a value to a variable with the desired width.
This has the unfortunate side-effect that every construction of a
WordX looks like:

    assign x = $unsigned(y);

In case of literals, this looks like:

    assign x = $unsigned(64'd5);

We will therefore never be able to use words as powerup values for
registers as HDL tools require a powerup value to not mention any
variables. The new blackbox function intercepts literals and will
generate appropriate literals as an expression:

    $unsigned(8'd5)

Thus adding the ability to use them as powerup values. A similar
approach will need to be added for other data constructors dealing
with literals.