Bundle: treat () as a sum type
() can be viewed as both a zero element product type (or 0-tuple),
and as a one constructor sum type.

In the past clash viewed () as a sum type:
   unbundle :: Signal dom () -> Signal dom ()
   bundle   :: Signal dom () -> Signal dom ()
Just like for example Bool.

This changed with: https://github.com/clash-lang/clash-prelude/issues/96
And since then clash views () as a product type:
   unbundle :: Signal dom () -> ()
   bundle   :: () -> Signal dom ()

But GHC-8.8 isn't fooled by the injectivity trickery to make this work,
so we'll have to revert it back to the sum type view.