Add a top-down AppProp transformation
The previous version was applied in a innerMost traversal,
propagating a single (innermost) argument at a time.

The new version tries to propagate as many arguments as possible,
and does it in a top-down traversal.

The goal is to ultimately reduces the number of transformation,
and traversals, ultimately reducing compile times.