Use non-recursive implementation of `rnfX` for `Vec`
Recursive function definitions can (in the general case) not be
translated by Clash. These recursive definitions can either be
eliminated by applying the right transformations, or by deducing that
the function isn't needed by the design in the first place. If it can't
do either, Clash will prevent netlist generation from running as it
risks getting stuck in an infinite loop. When a user defines a primitive
with /BlackBoxFunction/s, Clash can't predict what arguments will be
used. It will (rightfully!) reject cases where an argument uses a
recursive definition.

The definition of `rnfX` for `Vec` used to be implemented recursively,
causing a /BlackBoxFunction/ to potentially have an
`Undefined (Vec n a)` constraint. This is bad, as the class is used
quite ubiquitously. This commit therefore reimplements the definition
using a fold - a function Clash knows to be non recursive (not the bad
kind of recursive anyway :-)).