Evaluator: fix Signed pack/unpack
The following case would produce wrong results after being picked up by
the evaluator:

    unpack (resize (pack (-2 :: Signed 4)))

The evaluator stores literals as an `Integer`s, assuming that `pack`
would therefore be a no-op. In case of negative literals this isn't true
however. For example: `-2 :: Signed 4` its bit-representation is `1110`.
After packing/resizing to `Signed 8` it should now read `00001110`.
Clearly, this is a positive number, so leaving the original literal
integer untouched is wrong!