Change traversal strategy of `evalConst` from top-down to bottom-up
There is a bug in the evaluator where all data constructors are
considered lazy, even though their declaration says they have strict
fields. This causes some reductions to fail because the term under
the constructor is not in WHNF, which is what some of the evaluation
rules for certain primitive operations expect. Using a bottom-up
traversal works around this bug by ensuring that the values under
the constructor are in WHNF.

Using a bottomup traversal ensures that constants are reduced to NF,
even if constructors are lazy, thus ensuring more sensible/smaller
generated HDL.