Remove redundant imports
GHC-8.8 starts warning about what it call redundant imports:

	import Prelude
	import Data.Semigroup ((<>))

The second import will trigger a warning, because <> is already imported from Prelude.

But in older versions of base <> was not exported from Prelude,
so now we need to import it conditionally with CPP if we want to be warning free.

https://gitlab.haskell.org/ghc/ghc/wikis/migration/8.8#changes-to-unused-import-warnings